Working mutli-fx
New file structure, including backups folder for patches

pd_controller is now a class, with methods for interacting with pd, reloading patches, modifying the main patch to change the order in which the effects are sequenced. Metadata about the patch is now stored in the patch_meta.json file.

minor changes to gui.py, adapting to new PDController class.